backend/requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Settings
pydantic-settings==2.3.4

# Data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1

# RAG
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# LLM
groq==0.9.0

# Utils
python-dotenv==1.0.1

# Database
pymongo==4.7.3
bcrypt==4.1.3